requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
SQLAlchemy==2.0.4
alembic==1.9.4
pandas==1.5.3
numpy==1.24.2
requests==2.28.2
beautifulsoup4==4.11.2
pyyaml==6.0
python-dotenv==1.0.0
pytest==7.2.2
Werkzeug==2.2.3
Jinja2==3.1.2
APScheduler==3.10.0
gunicorn==20.1.0
schedule==1.2.0
click==8.1.3 